test: dissolve BusinessLogicTests into focused suites
Move tool, end-call, and audio-alignment coverage into Conversation
and event-handler tests so suite names match what they exercise.
